+++ Luna.Runtime/Functions/Lang/Range.cs
@@ -9,8 +9,8 @@ internal class Range : EmbeddedFunction
 {
     public override IRuntimeValue GetValue(ReadonlyArray<IRuntimeValue> argumentValues)
     {
-        var start = GetValueOrError<IntegerRuntimeValue>(argumentValues, 0).IntegerValue;
-        var count = GetValueOrError<IntegerRuntimeValue>(argumentValues, 1).IntegerValue;
+        var start = (int)GetValueOrError<IntegerRuntimeValue>(argumentValues, 0).IntegerValue;
+        var count = (int)GetValueOrError<IntegerRuntimeValue>(argumentValues, 1).IntegerValue;
         if (count < 0)
         {
             throw new RuntimeException("Count must be zero or greater.");
